# docs/swagger.yaml
basePath: /
definitions:
...
  models.GoogleAdGroupRequest:
    properties:
      campaign_resource_name:
        type: string
      name:
        type: string
    type: object
  models.GoogleAdRequest:
    properties:
      ad_group_resource_name:
        type: string
      description:
        type: string
      final_url:
        type: string
      headline:
        type: string
      name:
        type: string
    type: object
  models.GoogleCampaignRequest:
    properties:
      daily_budget_micros:
        type: integer
      name:
        type: string
    type: object
  models.GoogleKeywordDeleteRequest:
    properties:
      resource_name:
        type: string
    type: object
  models.GoogleKeywordRequest:
    properties:
      ad_group_resource_name:
        type: string
      cpc_bid_micros:
        type: integer
      keywords:
        items:
          type: string
        type: array
      match_type:
        description: EXACT, PHRASE, BROAD
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Multi-Platform Ads Management API (Meta & Google) for Realify.
  title: Realify API
  version: "1.0"
paths:
  /callback:
    get:
      description: Receives the auth code from Google and exchanges it for tokens
      parameters:
      - description: Auth Code from Google
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Google Ads OAuth Callback
      tags:
      - google_auth
  /google/adgroups:
    get:
      description: Fetch all ad groups for a specific campaign
      parameters:
      - description: Campaign Resource Name
        in: query
        name: campaign_resource_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: List Google Ad Groups
      tags:
      - google_adgroups
    post:
      consumes:
      - application/json
      description: Create a new ad group in a campaign
      parameters:
      - description: Ad Group Details
        in: body
        name: adgroup
        required: true
        schema:
          $ref: '#/definitions/models.GoogleAdGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Create Google Ad Group
      tags:
      - google_adgroups
  /google/adgroups/{id}/insights:
    get:
      description: Fetch real-time performance metrics for a specific ad group ID
      parameters:
      - description: Ad Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Get Google Ad Group Performance Insights
      tags:
      - google_adgroups
  /google/ads:
    get:
      description: Fetch all ads for a specific ad group
      parameters:
      - description: Ad Group Resource Name
        in: query
        name: ad_group_resource_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: List Google Ads
      tags:
      - google_ads
    post:
      consumes:
      - application/json
      description: Create a new responsive search ad
      parameters:
      - description: Ad Details
        in: body
        name: ad
        required: true
        schema:
          $ref: '#/definitions/models.GoogleAdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Create Google Ad
      tags:
      - google_ads
...
  /google/campaigns:
    delete:
      description: Remove a campaign by its resource name
      parameters:
      - description: Resource Name
        in: query
        name: resource_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Delete Google Campaign
      tags:
      - google_campaigns
    get:
      description: Fetch all campaigns for the configured customer ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: List Google Campaigns
      tags:
      - google_campaigns
    post:
      consumes:
      - application/json
      description: Create a new Google Ads campaign (SEARCH)
      parameters:
      - description: Campaign Details
        in: body
        name: campaign
        required: true
        schema:
          $ref: '#/definitions/models.GoogleCampaignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Create Google Campaign
      tags:
      - google_campaigns
  /google/campaigns/{id}/insights:
    get:
      description: Fetch real-time performance metrics for a specific campaign ID
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Get Google Campaign Performance Insights
      tags:
      - google_campaigns
...
  /google/keywords:
    delete:
      consumes:
      - application/json
      description: Remove a specific keyword by resource name
      parameters:
      - description: Delete Details
        in: body
        name: keyword
        required: true
        schema:
          $ref: '#/definitions/models.GoogleKeywordDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Delete Google Keyword
      tags:
      - google_keywords
    get:
      description: Fetch all keywords for an ad group
      parameters:
      - description: Ad Group Resource Name
        in: query
        name: ad_group_resource_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: List Google Keywords
      tags:
      - google_keywords
    post:
      consumes:
      - application/json
      description: Add keywords to an ad group
      parameters:
      - description: Keyword Details
        in: body
        name: keywords
        required: true
        schema:
          $ref: '#/definitions/models.GoogleKeywordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Create Google Keywords
      tags:
      - google_keywords
...
swagger: "2.0"